+++ app/assets/javascripts/factories/TryFactory.js
@@ -17,10 +17,6 @@
     var Try = function(data) {
       // This method converts the response from the API to angular objects.
       var self  = this;
-      if ( data.searchEngine !== undefined ) {
-        console.log('Data object creating Try has a searchEngine!');
-        console.log(data);
-      }
 
       if ( angular.isUndefined(data.search_engine) ) {
         console.log('We have an undefined data.search_engine so setting to Solr, should this ever happen?');
@@ -51,6 +47,7 @@
       self.curatorVars  = ngFriendlyCuratorVars;
 
       // Functions
+      self.formattedName   = formattedName;
       self.createFieldSpec = createFieldSpec;
       self.curatorVarsDict = curatorVarsDict;
       self.hasVar          = hasVar;
@@ -63,6 +60,15 @@
 
       // Bootstrap
       self.updateVars();
+
+      function formattedName() {
+        if (self.name.includes('Try ' + self.tryNo)){
+          return self.name;
+        }
+        else {
+          return self.name + ' - Try ' + self.tryNo;
+        }
+      }
 
       // Create a field spec from the string I'm
       // carrying around that stores that info