+++ test/controllers/api/v1/cases_controller_test.rb
@@ -145,10 +145,10 @@ module Api
             login_user matt
           end
 
-          test 'return a forbidden error' do
-            delete :destroy, params: { case_id: the_case.id }
+          test 'is perfectly okay, which is a different than before!' do
+            post :update, params: { case_id: the_case.id, archived: true }
 
-            assert_response :forbidden
+            assert_response :ok
           end
         end
 
@@ -159,11 +159,54 @@ module Api
             count_unarchived  = doug.cases.where(archived: false).count
             count_archived    = doug.cases.where(archived: true).count
 
-            delete :destroy, params: { case_id: one.id }
-            assert_response :no_content
+            put :update, params: { case_id: one.id, archived: true }
+            assert_response :ok
 
             assert_equal count_unarchived - 1,  doug.cases.where(archived: false).count
             assert_equal count_archived + 1,    doug.cases.where(archived: true).count
+          end
+        end
+
+        describe 'analytics' do
+          let(:one) { cases(:one) }
+
+          test 'posts event' do
+            expects_any_ga_event_call
+
+            perform_enqueued_jobs do
+              put :update, params: { case_id: one.id, archived: true }
+              assert_response :ok
+            end
+          end
+        end
+      end
+
+      describe 'Deleting a case' do
+        describe 'when it is the last/only case' do
+          let(:matt)      { users(:matt) }
+          let(:the_case)  { cases(:matt_case) }
+
+          before do
+            login_user matt
+          end
+
+          test 'is perfectly okay, it was old orthodoxy that every user has a case' do
+            delete :destroy, params: { case_id: the_case.id }
+            assert_response :no_content
+          end
+        end
+
+        describe 'when it is not the last/only case' do
+          let(:one) { cases(:one) }
+
+          test 'successfully deletes the case' do
+            count_cases = doug.cases.count
+
+            delete :destroy, params: { case_id: one.id }
+            assert_response :no_content
+
+            doug.cases.reload
+            assert_equal count_cases - 1, doug.cases.count
           end
         end
 
@@ -221,7 +264,7 @@ module Api
             count_unarchived  = doug.cases.where(archived: false).count
             count_archived    = doug.cases.where(archived: true).count
 
-            patch :update, params: { case_id: one.id, archived: false }
+            post :update, params: { case_id: one.id, archived: false }
             assert_response :ok
 
             one.reload
@@ -235,7 +278,7 @@ module Api
             count_unarchived  = doug.cases.where(archived: false).count
             count_archived    = doug.cases.where(archived: true).count
 
-            put :update, params: { case_id: one.id, archived: false }
+            post :update, params: { case_id: one.id, archived: false }
             assert_response :ok
 
             one.reload